# Conflicts:
#	support/cas-server-support-logging-config-cloudwatch/src/main/java/org/apereo/cas/logging/CloudWatchAppender.java

+++ support/cas-server-support-hazelcast-monitor/src/main/java/org/apereo/cas/monitor/HazelcastHealthIndicator.java
@@ -1,9 +1,9 @@
 package org.apereo.cas.monitor;
 
 import com.hazelcast.core.HazelcastInstance;
-import com.hazelcast.core.IMap;
-import com.hazelcast.instance.HazelcastInstanceProxy;
-import com.hazelcast.memory.MemoryStats;
+import com.hazelcast.instance.impl.HazelcastInstanceProxy;
+import com.hazelcast.internal.memory.MemoryStats;
+import com.hazelcast.map.IMap;
 import lombok.ToString;
 import lombok.extern.slf4j.Slf4j;
 import lombok.val;
@@ -87,7 +87,7 @@ public class HazelcastHealthIndicator extends AbstractCacheHealthIndicator {
 
         @Override
         public long getPercentFree() {
-            return (int) this.memoryStats.getFreeHeap() * PERCENTAGE_VALUE / this.memoryStats.getCommittedHeap();
+            return this.memoryStats.getFreeHeap() * PERCENTAGE_VALUE / this.memoryStats.getCommittedHeap();
         }
 
         @Override
